+++ test/unit/vitest.config.mjs
@@ -248,6 +248,9 @@ export default defineConfig( {
 				extends: true,
 				test: {
 					name: 'browser',
+					env: {
+						WP_TESTS_SKIP_STYLE_INJECTION: 'false',
+					},
 					include: vitestTests.browser,
 					setupFiles: [
 						path.join(